(refactor) Move `ShortcutKeyTest` to core package.
